Moved ui layout component into dedicated directory
