+++ electron/src/ui/layout/AppFooter.tsx
