+++ electron/src/ui/layout/AppHeader.tsx
